pokemon/requirements.txt
strands-agents-tools>=0.2.0
requests>=2.32.5